feat: move language from app/constants to app/translations
Update text color of subtitle to cater for dark mode
