--- app/constants/Language.ts
@@ -1,11 +1,0 @@
-export enum AppLanguage {
-  English = 'English',
-  German = 'German'
-}
-
-export function getAppLanguages (): string[] {
-  return [
-    AppLanguage.English,
-    AppLanguage.German
-  ]
-}
